+++ tests/testthat/test-utils_gt.R
@@ -79,7 +79,7 @@ test_that("add footnotes and headers - gt_group",{
 
 })
 
-test_that("add footnotes and headers - existing group info",{
+test_that("add footnotes and headers - existing subtitle info",{
   # create docorator object
   my_gt <- gt::exibble |>
     gt::gt(
@@ -126,6 +126,100 @@ test_that("add footnotes and headers - existing group info",{
 
 })
 
+test_that("add footnotes and headers - existing title info, no subtitle",{
+  # create docorator object
+  my_gt <- gt::exibble |>
+    gt::gt(
+      rowname_col = "row",
+      groupname_col = "group"
+    ) |>
+    gt::tab_header(title = "existing title")
+
+
+
+  docorator <- as_docorator(
+    x = my_gt,
+    header = fancyhead(
+      fancyrow(center = "first line header"),
+      fancyrow(left = "left header"),
+      fancyrow(center = "second line header"),
+      fancyrow(center = "third line header")
+    ),
+    footer = fancyfoot(
+      fancyrow(left = "footnote 1"),
+      fancyrow(left = "footnote 2"),
+      fancyrow(right = "timestamp")),
+    display_name = "my_first_gt",
+    save_object = FALSE
+
+  )
+
+  head_foot_gt <- hf_to_gt(docorator)
+
+  # subtitle
+  expect_equal(head_foot_gt$`_heading`$subtitle, gt::md("second line header<br>third line header<br>existing title"))
+
+  # no additional subtitles
+  docorator2 <- as_docorator(
+    x = my_gt,
+    display_name = "my_first_gt",
+    save_object = FALSE
+  )
+
+  head_foot_gt2 <- hf_to_gt(docorator2)
+
+  # subtitle
+  expect_equal(head_foot_gt2$`_heading`$subtitle, gt::md("existing title"))
+
+})
+
+test_that("add footnotes and headers - existing title + subtitle info",{
+  # create docorator object
+  my_gt <- gt::exibble |>
+    gt::gt(
+      rowname_col = "row",
+      groupname_col = "group"
+    ) |>
+    gt::tab_header(title = "existing title",subtitle = "TRT = Placebo, <br>X = Y")
+
+
+
+  docorator <- as_docorator(
+    x = my_gt,
+    header = fancyhead(
+      fancyrow(center = "first line header"),
+      fancyrow(left = "left header"),
+      fancyrow(center = "second line header"),
+      fancyrow(center = "third line header")
+    ),
+    footer = fancyfoot(
+      fancyrow(left = "footnote 1"),
+      fancyrow(left = "footnote 2"),
+      fancyrow(right = "timestamp")),
+    display_name = "my_first_gt",
+    save_object = FALSE
+
+  )
+
+  head_foot_gt <- hf_to_gt(docorator)
+
+  # subtitle
+  expect_equal(head_foot_gt$`_heading`$subtitle, gt::md("second line header<br>third line header<br>existing title<br>TRT = Placebo, <br>X = Y"))
+
+  # no additional subtitles
+  docorator2 <- as_docorator(
+    x = my_gt,
+    display_name = "my_first_gt",
+    save_object = FALSE
+  )
+
+  head_foot_gt2 <- hf_to_gt(docorator2)
+
+  # subtitle
+  expect_equal(head_foot_gt2$`_heading`$subtitle, gt::md("existing title<br>TRT = Placebo, <br>X = Y"))
+
+})
+
 
 test_that("no footnotes or headers",{
   my_gt <- gt::exibble |>
@@ -137,11 +231,11 @@ test_that("no footnotes or headers",{
 
   docorator <- as_docorator(
     x = my_gt,
+    header = NULL,
     footer = NULL,
     display_loc ="test",
     display_name = "my_first_gt",
     save_object = FALSE
-
   )
 
   head_foot_gt <- hf_to_gt(docorator)
@@ -204,4 +298,67 @@ test_that("apply_to_grp works",{
 
   # check apply_to_grp works for gt_tbl
   expect_identical(options_tbl, apply_to_grp(func, arg_list_tbl))
+})
+
+test_that("Create png from ggplot", {
+
+  withr::with_tempdir({
+    gg <- ggplot2::ggplot(mtcars) +
+      ggplot2::aes(x = disp, y = mpg) +
+      ggplot2::geom_point()
+
+
+    image_paths <- gg_to_image(gg, path = getwd())
+
+    expect_length(image_paths, 1)
+
+      expect_equal(
+        file.exists(image_paths),
+        c(TRUE)
+      )
+  })
+
+})
+
+test_that("Create set of png from list of ggplots", {
+
+  withr::with_tempdir({
+    gg1 <- ggplot2::ggplot(mtcars) +
+      ggplot2::aes(x = disp, y = mpg) +
+      ggplot2::geom_point()
+
+
+    gg2 <- ggplot2::ggplot(mtcars) +
+      ggplot2::aes(x = hp, y = mpg) +
+      ggplot2::geom_point()
+
+    image_paths <- gg_to_image(list(gg1, gg2), path = getwd())
+
+    expect_length(image_paths, 2)
+
+    expect_equal(
+      file.exists(image_paths),
+      c(TRUE, TRUE)
+    )
+  })
+})
+
+test_that("Extract header footer information from ggplot", {
+
+  ggplot1 <- ggplot2::ggplot(data = mtcars, ggplot2::aes(y=cyl, x=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(title = "title1", subtitle = "subtitle1", alt = "alt", tag = "tag1", caption = "footnote1")
+
+  stripped_ggplot <- ggplot2::ggplot(data = mtcars, ggplot2::aes(y=cyl, x=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(alt = "alt")
+
+  hf_ggplot <- hf_strip(ggplot1)
+
+  expect_equal(stripped_ggplot$labels, hf_ggplot$display$labels)
+
+  expect_equal(hf_ggplot$head_data, "title1")
+  expect_equal(hf_ggplot$subhead_data, c("subtitle1", "tag1"))
+  expect_equal(hf_ggplot$foot_data, "footnote1")
+
 })